Replicating without decoding
This commit introduces a binary that can logically replicate (create a
subscription, as in the psql CREATE SUBSCRIPTION command) and spews the
wal data we receive from each update.

Next steps are to implement a logical decoder that can parse the
pgoutput format. Will look at pgx to see how it parses the Postgres
protocol for examples of how best to approach this.