Better separation of pipeline stages
Ignore importing and serialization for now, the aim of this change is to
provide a stream of Go native types in the modification and schema
output.

To clarify: schemas are always going to be in Avro format, but the
modifications may or may not be Avro encoded. I'm yet to decide if the
entire message should be Avro, or just the Before/After elements.