Uniquely identify publications
We need a way to identify if a publication gets recreated under a new
name, as this would signify that all previous imports are now invalid.
An entirely new table seems overkill, and would require some
intelligence to discover which publication it currently references (is
it the one with the max(created_at)?).

Instead, comment on the publication in postgres. If a publication is
ever dropped, the comment will also be lost, making it easy to detect
whenever we require a fresh import of the tables.