Track imports against subscription
Publications can have multiple subscriptions. We only want to schedule
an import if a table is in the publication without an associated import
for the same publication/subscription pair.

This change adds the subscription_name column to import_jobs, and
modifies the importer to ensure we track import progress against
subscriptions. This incompletely addresses the problem, as tables
removed and then re-added to the publication won't retrigger an import,
and if a replication slot is removed and re-added we also won't detect
that. This is ok for now.