pg2sink resync --table example.public
Don't remove a table from the publication, but expire all the import
jobs so that we re-trigger an import. This is useful when a specific
subscription needs reimporting, and you don't want to remove the table
from the publication for all dependent subscriptions.

It can also be useful if a sink requires resyncing from scratch.