BigQuery schema sink resource
Implement syncing of schemas into a BigQuery dataset. For now we panic
whenever a modification comes down the stream, but the schemas appear to
be created successfully.

Changelog entries are intended to be placed in a table with _raw suffix,
while a view presents that _raw data with the most-recent rows only.