Always close messages in subscription
We should close the messages channel whenever the StartReplication
function returns. The way the subscription interface works is a bit
rubbish, as it doesn't express temporal couplings using the type system,
so it's unclear that StartReplication can be called just once.

This unsafe interface is, however, more sane if it takes care of the
closing of this channel, as the alternative is to accidentally hang
indefinitely.