SchemaHandler doesn't route
A SchemaHandler shouldn't have to worry about how we'll route
modification to this inserter. The contract is that we'll return an
Inserter that is valid for modifications matching this schema. We can
safely hardcode that we use the schema Namespace as a routing key into
our generic sink implementation.

This commit removes the RouteMatchAll value that was previously used to
tell the Router to map all incoming modifications to the same inserter.
We'll now have many routes pointing at the same inserter, but that's
probably ok.