pkg/subscription uses db.SQL
db.SQL is the highest level of abstraction over Postgres connections. It
makes sense for everything to consume this, and optionally drop into a
pgx connection where necessary.

This commit is the first that changes all the interfaces into db.SQL,
with the only downside being the interfaces for querying are slightly
more dangerous.