pkg/subscription: confirmed_flush_lsn
This commit uses the confirmed_flush_lsn field of pg_replication_slots
to resume replication from the last LSN that we confirmed receipt for
with the upstream server. This is what you'd expect from a replicating
process, and until now we had either skipped changes or started from
scratch.

---

Prior to refactoring the subscription behaviour, we used to request WAL
from LSN(0) from the upstream server. This caused us to receive all
changes since the start of time (or whenever the server has WAL until)
which I'd noticed but hadn't debugged.

After refactoring, I'd used the WAL position given by IDENTIFY SYSTEM to
resume replication. As this was the most recent WAL position, we'd
effectively skip any changes that happened between our previous life and
now.