cmd/pg2sink/cmd: better shutdown
Bring code into the cmd package, so we separate it from main. This
allows us to import it, and is the start of splitting each command into
different files.

While here, improve our concurrency control so each long-lived
components can respond to a shutdown request. This prevents us from
exiting with an error if we were asked to shutdown, and did so
peacefully.