imports: functional skip locked worker
Bring back a database queue in the form of imports.Worker. We could
generalise this module into a queue package, and either duck type our
job tables or use a centralised queue table for tracking outstanding
work.

Either of these seems premature. I don't anticipate another use case for
shared concurrency in this project, and it's simpler for workers to
compete over the data rows that represent imports.