Respect context in file inserter
When using the BigQuery sink, we don't get any data because the context
keep expiring. By ignoring this context in the file inserter, we miss
that something upstream is incorrectly cancelling the context we're
using to perform the insert.

Best respect it, even if the file write doesn't strictly need a context.