pkg/types.Decoder
This project needs to care a lot about types, and how we handle
different fields from Postgres. Due to the way it was built, the code
aimed to produce a sensible intermediate format (pkg/changelog) that
had parsed Postgres types into Golang native types.

While this worked, it meant we decided how to parse Postgres fields
before we ever knew where we'd put that data. There was a small effort
to abstract serialization on-top of the changelog package, but it was
clear that wouldn't be the end of the story even at the start.

Now we have first-class support for arbitrary sinks, it becomes clear
that we can't make serialization decisions without knowledge of where we
aim to put the data. Sure, we could translate things to a feature-rich
intermediary and then bail if the downstream doesn't support it, but
that sucks too.

The best idea is to push decoding/serialization decisions into the
sinks. In a full pgsink stream, we expect:

    postgres (import & stream)
      -> intermediary
      -> sink

If we can have sinks advertise the PG types they support and how they
want them parsed up-front, we can provide that to the Postgres import
and stream components and have them make appropriate decisions based on
the sink.

That is what this commit does. It established a (very inefficient)
decoder that maps Postgres OID to the chosen scanner, and the empty
Golang native type the scanner will produce. Each sink bundles a decoder
to ensure the Postgres upstream parses content that is compatible with
the sink.

In future, we'll want to tidy the abstractions so they fit more neatly
into package responsibilites (sinks should probably all provide a
decoder? and that be exposed in the interface?) but we also force user
decisions, such as:

1. Should pgsink reject Postgres tables with unsupported types?
2. Or should we fallback to sending the Postgres textual representation
   if we don't have a supported type?

People will have different preferences depending on the sink and their
use case, so it needs to be configurable.

This commit sets the stage for all of this, while making a massive mess.