Stash logger in context, for db logs
Stash a logger in trace contexts, so any traced context can provide
downstream code with a logger. This allows us to change the database
logger so it uses a logger associated with the current trace, rather
than the global one, which was making it impossible to link database
traces to the originating operation.