Exponentially backoff when import jobs fail
Repeating the job over and over with no repeat was a bit aggressive.
This allows us to gradually back-off.

It's implemented with jet expressions, and pushing a lot of work into
the database. This is a bit complicated but not terrible- the
alternative is to calculate the next due at time when a job errors,
which might be nicer. Not least because it materializes that value into
the database, for debugging.