Embed Javascript assets into binary
pgsink is designed to be shipped as a single binary. It might be
distributed in a docker container, but it's really nice if a user can
install the binary and that's it, everything works.

Until we had a web component, this was true. But now we have a web
control panel, we were shipping a `serve` command that lacked the assets
it needed to serve.

By using the go-1.16 release candidate, we can use the embed directive
to load web assets directly into the binary. This is a huge usability
improvement for the cost of the embedded asset size, which is a small
fraction (~10%) of the static binary:

    # web/assets directory, about 2.2M
    $ gdu -sb web/build
    2231689 web/build

    # Go 1.16, with embed
    $ go build -ldflags "-s -w -X main.Version=40437a6-dev" -o bin/pgsink cmd/pgsink/*.go
    $ du -sh bin/pgsink
    24M (25326560)

    # Go 1.16, with empty embedded files
    $ go build -ldflags "-s -w -X main.Version=40437a6-dev" -o bin/pgsink cmd/pgsink/*.go
    $ du -sh bin/pgsink
    22M (23094240)

It is worth noting that I tend to use `-s -w` flags to reduce binary
size, which drops the symbol tables. I don't know if this will affect
the embed, though it doesn't appear to.