BigQuery: support table clustering
Support clustering the raw tables by the Postgres primary key, which
should make queries a lot more efficient.

While here, fix a clear bug in the typing of the mapping. I'd been
relying on the tests to verify this, but they weren't complete. Clear
evidence we need an integration test!