127.0.0.1 for testing
I wondered why localhost wasn't working, and only realised I hadn't
closed the listener when fixing the tests in CI.

On Darwin, something about hitting the firewall when binding to the port
meant we'd recover from the error and the test would succeed. Not ideal
really, this is much healthier (and avoids the annoying firewall warning
on each test).